The pool screen now shows all the brackets in the pool.
This took a lot of CSS hacking, to get the ng-grid bracket too look right, when ng-included into the pool page. Also changed the way bracket ids are stored in the pool, to be keyed by the user id. This makes it easy to tell which bracket belongs to the currently logged in user.

I also downloaded all of the rawgithub.com static assets, and checked them into a thirdParty directory for CSS and JS.