Trying to get pool and bracket views working on pick100pool.firebaseapp.com.
firebaseapp.com does not like ng-cloak, and $scope.bracketIds in poolsControllers was never resolving, so I just ripped it out and switched to use pool.brackets in the view. This simplified the controller, at the cost of moving the "edit your bracket" button into the bracket view, and not being able able to filter the user's own bracket from the list.

I also reverted my changes to simpleLoginTools.js, in hopes that would fix the un-cloaking problems.

The create bracket button is still not showing up for logged out users, which is a bug but not the end of the world.