Changed back-end to pull round 1 play-in games
This is so that the teams in the play-in games will be available in the UI to choose for a bracket, before the play-in games have finished. Round 1 (i.e. play-in games) are scored as zero points for the winner.

Also changed web.js to check for updated game scores every 5 minutes instead of 1 hour.