Prevent user from creating a pool or bracket if they're not logged in
Also changed the "create a new pool" link to be a button. Only show the buttons for creating a pool and bracket if the user is logged in; this prevents the broken login/register page from being displayed when the user is not authenticated and tries to do an action which requires them to be logged in.