fixes #19 performance improvements to bracket-view.js
Download all the teams at once from firebase, instead of downloading each team separately for the bracket. It looks like $scope.teamsWithScores might be shared across brackets on the same page, so I added checks to make sure the point totals for each team are only calculated once. The downside of this change is that I don't think the page will automatically update anymore when games complete, and points for new rounds are added to the team in firebase. This is an unfortunate tradeoff that's necessary to make the pool page useable when there are more than a couple of brackets.

Also, I added a $rootScope decorator to unsubscribe all watchers when an object in the scope is destroyed.