fixed bug where brackets are re-added to the list instead of updated
Changed how individual brackets are downloaded from firebase, and populated in the allBracketsInPool array. This new approach is quick to initially populate the brackets, but still supports updating brackets on the fly when the bracket fields (e.g. points or teams remaining) changes.

Also replaced calls to $on('value', ...) with $getRef().once('value', ...) in the bracket create/edit and view controllers, because it seems a lot faster and does the same thing.