navbar header now shows Play Now link the same as the Rules link
This makes it easier for users on small screens to click Play Now; they no longer have to use the dropdown menu, which doesn't disappear after you click an entry. This also increases the font size of the Play Now link to be the same as the Rules link, which reflects the equal importance that both pages have.