Don't wait for auth; bind currentUserBracketId if $scope.auth.user is set
This attempts to fix a bug where the "Select Your Teams" button wouldn't show up, even if the user was logged in. The problem was that the waitForAuth callbacks were never getting called, so the next function which sets $scope.model.currentUserHasLoaded and binds $scope.model.currentUserBracketId was never getting called. Refreshing the page always seemed to work, though. I also moved the hasTourneyStarted check to after that block, to give the waitForAuth method slightly more time.