Adds Change Password page and Reset Password link to header
I removed the extra stuff so now account.html and AccountCtrl in controllers.js just supports changing the user's password. If we ever want to add support back in for changing the user's name that's easy, but I don't recommend supporting changing their email address since that's implemented by lots of steps in firebase.

I also removed the security filter that redirects to the login.html page if the request is to a secure page and the user isn't logged in, since that page isn't operational. Easy enough to add this back to module.routeSecurity.js when/if I build a login page that's separate from the header.