Removes fetchAllTeams() in favor of loading them in just one place
No need to make allTeamsSnapshot a global var, when it's only used in one place: updatePointsForBracket(). What I noticed was happening is that this would only fail when run in node, and only on the second try. So I think the is_eliminated field got nulled out somehow in the firebase client's local cache.