fixes set vs. transaction conflict bug in adding team to firebase if new
Using Firebase#transaction for this, but Firebase#set and #update for changing values caused the local client's cache to get out of sync with the server, which caused the client's snapshot view to NOT contain the is_eliminated=true property, probably because it wasn't included in the Firebase#transaction call. I rewrote the code to add if the team to firebase if it doesn't already exist, using Snapshot#exists and Firebase#set.