upgrades ng-grid to ui-grid 3.2.9, but fails to fix slow bracket-view grid
This does *not* use the "angular-ui-grid" bower package, because its semver requires angular 1.4+, and this app is using 1.2. I downloaded the ui-grid JS and CSS directly from https://github.com/angular-ui/bower-ui-grid/tree/v3.2.9.

In general the ui-grid 3.x APIs have gotten more intuitive and flexible, without requiring as much overrides of built-in templates.

This didn't seem to help much in speeding up the grid in bracket-view, so I'm abandoning this work before everything is converted and working perfectly. Here are a few things left to do, if this upgrade is ever pursued:

* replace ng-grid-flexible-height.js with some other way to adjust bracket list height to fit rows
* broken: sorting by points col in bracket list
* fix ng* CSS classes in app.css, which are hard-coded to ngGrid classes
* is bracket-list.js logic to handle bracket data changing still needed?
* is ui-grid-resize-columns needed?