Concatenates the team.short_name and team.id fields, so it's unique
This fixes the problem where San Diego State and South Dakota State both share the same short_name and abbreviation fields. I tried using the medium_name or full_name instead, but that contains a period in the case of "N.C. State" (even for the full_name field!), which is not an allowed character for firebase paths. This is a backwards incompatible change -- can't run this code with previous year's data. :(

Also improved error handing by adding try/catch blocks around code that's called in a promise.